Fixed possible integer overflow in Fat32BootSector.writeCopy(..).

+++ src/main/java/de/waldheinz/fs/fat/Fat32BootSector.java
@@ -190,7 +190,7 @@ final class Fat32BootSector extends BootSector {
      */
     public void writeCopy(BlockDevice device) throws IOException {
         if (getBootSectorCopySector() > 0) {
-            final long offset = getBootSectorCopySector() * SIZE;
+            final long offset = (long)getBootSectorCopySector() * SIZE;
             buffer.rewind();
             buffer.limit(buffer.capacity());
             device.write(offset, buffer);
